+++ seurust/tests/testthat.R
@@ -1,1 +1,4 @@
 library(testthat)
+library(seurust)
+
+test_check("seurust")